Un monton de cambios

+++ ZonaRTX/Models/DetallesPedidosModel.cs
@@ -6,7 +6,7 @@
         public int id_pedido { get; set; }
         public int id_producto { get; set; }
         public int cantidad { get; set; }
-        public float precio_unitario { get; set; }
+        public decimal precio_unitario { get; set; }
 
     }
 }
